+++ package.json
@@ -9,8 +9,7 @@
     "test": "wsrun -m test",
     "dev": "wsrun -p '*-end' -m dev",
     "build": "wsrun -p '*-end' -m build",
-    "start": "wsrun -p '*-end' -m start",
-    "cron": "yarn workspace back-end cron"
+    "start": "wsrun -p '*-end' -m start"
   },
   "workspaces": [
     "packages/*"
