+++ packages/back-end/package.json
@@ -12,7 +12,6 @@
     "start": "node dist/server.js",
     "test": "jest --forceExit --coverage --verbose --detectOpenHandles",
     "type-check": "tsc --pretty --noEmit",
-    "cron": "node dist/cron.js",
     "generate-dummy-data": "node --stack-size=8192 ./test/data-generator/data-generator.js",
     "import-dummy-data": "node --stack-size=8192 ./test/data-generator/import.js"
   },
